move Appointment tests into a different file
